feat: cohort creation for the cancer project
