+++ psycop/common/feature_generation/loaders/raw/load_cancer_outcomes.py
@@ -1,63 +1,26 @@
-"""Loaders for cancer outcomes."""
-#from __future__ import annotations
+"""Loaders that are specific to the cancer project."""
+# pylint: disable=missing-function-docstring
+
+from pathlib import Path
+
 import pandas as pd
 
-from psycop.common.feature_generation.loaders.raw.sql_load import sql_load
+from psycop.common.feature_generation.utils import data_loaders
+
+CANCER_DATA_DIR = Path(r"E:\shared_resources") / "cancer"
 
 
-# LPR3, both in and outpatient
-df_lpr3_preproc = sql_load(
-    "SELECT * FROM [fct].FOR_LPR3kontakter_psyk_somatik_inkl_2021",
-)[["dw_ek_borger", "datotid_lpr3kontaktstart", "adiagnosekode", "shakkode_lpr3kontaktansvarlig"]]
+@data_loaders.register("any_cancer")
+def any_cancer() -> pd.DataFrame:
+    """Loads the outcome variable for the cancer prediction project.
+    See `outcome_specification/outcome_spec.py` for details.
+    """
+    df = pd.read_csv(CANCER_DATA_DIR / "cancer_cohort.csv")
+    df.drop(columns="Unnamed: 0", inplace=True)
+    df = df.rename(columns={"datotid_start": "timestamp"})
+    df["timestamp"] = pd.to_datetime(df["timestamp"]).apply(
+        lambda x: x.replace(tzinfo=None),
+    )
+    df["value"] = 1
 
-df_lpr3_preproc.rename(columns={
-            "datotid_lpr3kontaktstart": "datotid_start",
-            "shakkode_lpr3kontaktansvarlig": "shakafskode",
-            }, inplace = True
-        )
-
-
-# LPR2
-#inpatient
-df_lpr2_inp_preproc = sql_load(
-    "SELECT * FROM [fct].FOR_indlaeggelser_psyk_somatik_LPR2_inkl_2021",
-)[["dw_ek_borger", "adiagnosekode", "datotid_indlaeggelse", "shakKode_kontaktansvarlig"]]
-
-df_lpr2_inp_preproc.rename(columns={
-            "datotid_indlaeggelse": "datotid_start",
-            "shakKode_kontaktansvarlig": "shakafskode",
-            }, inplace = True
-        )
-
-#outpatient
-df_lpr2_outp_preproc = sql_load(
-    "SELECT * FROM [fct].FOR_besoeg_psyk_somatik_LPR2_inkl_2021",
-)[["dw_ek_borger", "diagnoseKode", "datotid_start", "shakafskode"]]
-
-df_lpr2_outp_preproc.rename(columns={
-            "diagnoseKode": "adiagnosekode",
-            }, inplace = True
-        )
-
-
-# Combine all
-all_visits_combined = pd.concat([df_lpr3_preproc, df_lpr2_inp_preproc, df_lpr2_outp_preproc], ignore_index=True)
-
-#extract first visit to psych in RM
-df_first_psych_visit = all_visits_combined[all_visits_combined["shakafskode"].str.startswith("6600")].groupby(["dw_ek_borger"])["datotid_start"].min().to_frame().reset_index()
-df_first_psych_visit.rename(columns={"datotid_start": "datotid_first_psych_visit"}, inplace = True)
-
-
-# Extract cancer diagnosis
-DIAGNOSIS_CODE = "DC"
-
-df_cancer_visits = all_visits_combined[all_visits_combined["adiagnosekode"].str.startswith(DIAGNOSIS_CODE, na=False)]
-
-#only include patient that have been diagnosed with cancer after their first visit to psychiatry and after 2013
-df_cancer_visits_ = df_cancer_visits.merge(df_first_psych_visit, on="dw_ek_borger")
-
-df_cancer_visits_after_first_psych_visit = df_cancer_visits_[(df_cancer_visits_["datotid_start"] > df_cancer_visits_["datotid_first_psych_visit"]) & (df_cancer_visits_["datotid_start"] > "2013-01-01")]
-
-
-# Extract only the first visit
-df_cancer_visits_after_first_psych_visit_only_first = df_cancer_visits_after_first_psych_visit.loc[df_cancer_visits_after_first_psych_visit.groupby(["dw_ek_borger"])["datotid_start"].idxmin()].reset_index(drop=True)
+    return df.reset_index(drop=True)
