+++ psycop/projects/cancer/feature_generation/cohort_definition/eligible_prediction_times/cancer_eligible_config.py
