+++ psycop/projects/cancer/feature_generation/cohort_definition/outcome_specification/first_cancer_diagnosis.py
@@ -1,11 +1,30 @@
 import pandas as pd
 import polars as pl
 
-def get_first_cancer_diagnosis():
-    pass
+from psycop.common.feature_generation.loaders.raw.load_cancer_outcomes import (
+    any_cancer,
+)
 
-def get_type_of_first_cancer_diagnosis():
-    pass
 
-def get_time_of_cancer_diagnosis():
-    pass
+def get_first_cancer_diagnosis() -> pd.DataFrame:
+    df = any_cancer()
+
+    df_first_cancer_diagnosis = (
+        df.sort_values("timestamp", ascending=True)
+        .groupby("dw_ek_borger")
+        .first()
+        .reset_index(drop=False)
+    )
+    return df_first_cancer_diagnosis[["dw_ek_borger", "timestamp", "value"]]
+
+def get_type_of_first_cancer_diagnosis() -> pd.DataFrame:
+    df = any_cancer()
+
+    df_first_cancer_diagnosis = (
+        df.sort_values("timestamp", ascending=True)
+        .groupby("dw_ek_borger")
+        .first()
+        .reset_index(drop=False)
+    )
+
+    return df_first_cancer_diagnosis[["dw_ek_borger", "adiagnosekode"]]
